feat: adding running tab to recipe details
